fix: move mock data and apply hock convention
